# web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Food & Health Planner",
  "description": "Privater Essensplan und Einkaufsliste für 2 Personen.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/ingredients": {
   "get": {
    "tags": [
     "Zutaten"
    ],
    "summary": "Zutaten suchen",
    "operationId": "list_ingredients_api_ingredients_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "description": "Suchtext (Teilstring und Ähnlichkeit)",
       "title": "Q"
      },
      "description": "Suchtext (Teilstring und Ähnlichkeit)"
     },
     {
      "name": "source",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "bls",
          "off",
          "manual"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Nur Zutaten dieser Quelle",
       "title": "Source"
      },
      "description": "Nur Zutaten dieser Quelle"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 25,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "include_hidden",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Ausgeblendete Zutaten mit anzeigen",
       "default": false,
       "title": "Include Hidden"
      },
      "description": "Ausgeblendete Zutaten mit anzeigen"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IngredientListOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "Zutaten"
    ],
    "summary": "Eigene Zutat anlegen",
    "operationId": "create_ingredient_api_ingredients_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IngredientIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IngredientDetailOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/ingredients/{ingredient_id}": {
   "get": {
    "tags": [
     "Zutaten"
    ],
    "summary": "Zutat mit Synonymen",
    "operationId": "get_ingredient_api_ingredients__ingredient_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "ingredient_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Ingredient Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IngredientDetailOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Zutaten"
    ],
    "summary": "Zutat ändern",
    "operationId": "patch_ingredient_api_ingredients__ingredient_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "ingredient_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Ingredient Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IngredientPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IngredientDetailOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Zutaten"
    ],
    "summary": "Eigene Zutat löschen",
    "operationId": "delete_ingredient_api_ingredients__ingredient_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "ingredient_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Ingredient Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/ingredients/{ingredient_id}/synonyms": {
   "post": {
    "tags": [
     "Zutaten"
    ],
    "summary": "Synonym hinzufügen",
    "operationId": "add_synonym_api_ingredients__ingredient_id__synonyms_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "ingredient_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Ingredient Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SynonymIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IngredientDetailOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/ingredients/{ingredient_id}/synonyms/{alias}": {
   "delete": {
    "tags": [
     "Zutaten"
    ],
    "summary": "Synonym entfernen",
    "operationId": "remove_synonym_api_ingredients__ingredient_id__synonyms__alias__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "ingredient_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Ingredient Id"
      }
     },
     {
      "name": "alias",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Alias"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "IngredientDetailOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "source_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Code"
     },
     "nutrients": {
      "$ref": "#/components/schemas/NutrientsPer100Out"
     },
     "density_g_per_ml": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Density G Per Ml"
     },
     "piece_g": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Piece G"
     },
     "is_fish": {
      "type": "boolean",
      "title": "Is Fish"
     },
     "is_potassium_salt": {
      "type": "boolean",
      "title": "Is Potassium Salt"
     },
     "shelf_days": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shelf Days"
     },
     "hidden": {
      "type": "boolean",
      "title": "Hidden"
     },
     "synonyms": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Synonyms"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "category",
     "source",
     "source_code",
     "nutrients",
     "density_g_per_ml",
     "piece_g",
     "is_fish",
     "is_potassium_salt",
     "shelf_days",
     "hidden",
     "synonyms"
    ],
    "title": "IngredientDetailOut"
   },
   "IngredientIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Name"
     },
     "category": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 60
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "kcal_100": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 900.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Kcal 100"
     },
     "protein_100": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 100.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Protein 100"
     },
     "fat_100": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 100.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Fat 100"
     },
     "carb_100": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 100.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Carb 100"
     },
     "fiber_100": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 100.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Fiber 100"
     },
     "salt_100": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 100.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Salt 100",
      "description": "Salz in g je 100 g"
     },
     "micros": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Micros",
      "description": "Schlüssel wie `zinc_mg`, je 100 g"
     },
     "density_g_per_ml": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 20.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Density G Per Ml"
     },
     "piece_g": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 5000.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Piece G",
      "description": "Gewicht eines Stücks in g"
     },
     "is_fish": {
      "type": "boolean",
      "title": "Is Fish",
      "default": false
     },
     "is_potassium_salt": {
      "type": "boolean",
      "title": "Is Potassium Salt",
      "default": false
     },
     "shelf_days": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 3650.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Shelf Days"
     },
     "hidden": {
      "type": "boolean",
      "title": "Hidden",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "IngredientIn",
    "description": "Eine manuell gepflegte Zutat (Werte je 100 g)."
   },
   "IngredientListOut": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/IngredientOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "limit": {
      "type": "integer",
      "title": "Limit"
     },
     "offset": {
      "type": "integer",
      "title": "Offset"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "limit",
     "offset"
    ],
    "title": "IngredientListOut"
   },
   "IngredientOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "source_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Code"
     },
     "nutrients": {
      "$ref": "#/components/schemas/NutrientsPer100Out"
     },
     "density_g_per_ml": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Density G Per Ml"
     },
     "piece_g": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Piece G"
     },
     "is_fish": {
      "type": "boolean",
      "title": "Is Fish"
     },
     "is_potassium_salt": {
      "type": "boolean",
      "title": "Is Potassium Salt"
     },
     "shelf_days": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shelf Days"
     },
     "hidden": {
      "type": "boolean",
      "title": "Hidden"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "category",
     "source",
     "source_code",
     "nutrients",
     "density_g_per_ml",
     "piece_g",
     "is_fish",
     "is_potassium_salt",
     "shelf_days",
     "hidden"
    ],
    "title": "IngredientOut"
   },
   "IngredientPatch": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "kcal_100": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Kcal 100"
     },
     "protein_100": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Protein 100"
     },
     "fat_100": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fat 100"
     },
     "carb_100": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Carb 100"
     },
     "fiber_100": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fiber 100"
     },
     "salt_100": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salt 100"
     },
     "micros": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "number"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Micros"
     },
     "density_g_per_ml": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Density G Per Ml"
     },
     "piece_g": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Piece G"
     },
     "is_fish": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Fish"
     },
     "is_potassium_salt": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Potassium Salt"
     },
     "shelf_days": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shelf Days"
     },
     "hidden": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hidden"
     }
    },
    "type": "object",
    "title": "IngredientPatch",
    "description": "Teiländerung; nur gesendete Felder gelten."
   },
...
   "NutrientsPer100Out": {
    "properties": {
     "kcal": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Kcal"
     },
     "protein_g": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Protein G"
     },
     "fat_g": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fat G"
     },
     "carb_g": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Carb G"
     },
     "fiber_g": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fiber G"
     },
     "salt_g": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salt G"
     },
     "micros": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Micros"
     }
    },
    "type": "object",
    "required": [
     "kcal",
     "protein_g",
     "fat_g",
     "carb_g",
     "fiber_g",
     "salt_g",
     "micros"
    ],
    "title": "NutrientsPer100Out"
   },
...
   "SynonymIn": {
    "properties": {
     "alias": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Alias"
     }
    },
    "type": "object",
    "required": [
     "alias"
    ],
    "title": "SynonymIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "description": "Zugangstoken der Person (`python -m app.cli`)",
    "scheme": "bearer"
   }
  }
 }
}